Add files via upload
I have made a bash script that will scrape the API for the NIST CVE Vulnerability Index (National Vulnerability Database) and calculate the average impact score of CVE's for the selected timeframe. The script then formats the data from the API into an excel spreadsheet. As a security measure, this script requires the input of the correct password to be able to access it. The password is encrypted with a modern hashing algorithm (SHA256).